"fix" unnecessary error upon loading an image in tui image editor, also restructure editor paths and names
