+++ projects/bp-gallery/src/components/common/editors/helpers/format-texts.ts
