+++ projects/bp-gallery/src/components/common/editors/helpers/jodit-config.ts
