+++ projects/bp-gallery/src/components/views/start/DailyPicture.tsx
@@ -10,7 +10,7 @@ import PictureView from '../picture/PictureView';
 import { Portal } from '@mui/material';
 import { useTranslation } from 'react-i18next';
 import { formatTimeStamp } from '../../../helpers/format-timestamp';
-import Editor from '../../common/editor/Editor';
+import TextEditor from '../../common/editors/TextEditor';
 import { FolderSpecial, Event } from '@mui/icons-material';
 import { asApiPath } from '../../../helpers/app-helpers';
 
@@ -86,7 +86,7 @@ const DailyPicture = () => {
             <div className={'flex flex-col max-w-4xl'}>
               <p className={'line-clamp-5'}>
                 <h4 className={'text-lg my-1'}>{t('pictureFields.descriptions')}:</h4>
-                <Editor value={description} />
+                <TextEditor value={description} />
               </p>
               <div className={'flex-1'} />
               <div className={'flex items-center gap-2 my-2'}>
